requirements.txt
Flask==3.0.0
Flask-PyMongo==3.0.1
pymongo[srv]==4.11.3
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
google-generativeai==0.3.2
APScheduler==3.10.4
Werkzeug==3.0.1
email-validator==2.1.0.post1
WTForms==3.1.1
dnspython==2.7.0
alembic==1.13.1
cryptography==42.0.5
tenacity==8.2.3 